+++ src/components/panel/SettingsPanel.tsx
@@ -31,8 +31,8 @@ import Switch from '../ui/Switch';
 import Input from '../ui/Input';
 import Slider from '../ui/Slider';
 import { ThemeProps, THEMES, DEFAULT_THEME_ID } from '../../utils/themes';
-import { Invokes, KEYBINDING_DEFINITIONS, KEYBINDING_SECTIONS, KeybindingDefinition } from '../ui/AppProperties';
-import { arraysEqual, codeToDisplayLabel, formatKeyCode, normalizeCombo } from '../../utils/keyboardUtils';
+import { Invokes } from '../ui/AppProperties';
+import { arraysEqual, codeToDisplayLabel, formatKeyCode, KeybindingDefinition, KEYBINDING_DEFINITIONS, KEYBINDING_SECTIONS, normalizeCombo } from '../../utils/keyboardUtils';
 import Text from '../ui/Text';
 import { TextColors, TextVariants, TextWeights } from '../../types/typography';
 import { useOsPlatform } from '../../hooks/useOsPlatform';
